+++ src/main/java/com/umc/halo/domain/member/repository/MemberRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.*;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.*;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -34,4 +35,5 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
        """)
     Optional<Member> findByIdForUpdate(@Param("id") Long id);
 
+    List<Long> findAllIds();
 }
